# artifacts/api-server/src/routes/profile.ts
import {
  Router,
  type IRouter,
} from "express";
import { randomBytes } from "node:crypto";
import {
  db,
  cvDocumentsTable,
  profilesTable,
  professionChangePaymentsTable,
  premiumPaymentsTable,
  savedJobsTable,
  pushDevicesTable,
} from "@workspace/db";
import { and, desc, eq } from "drizzle-orm";
import { UpsertProfileBody } from "@workspace/api-zod";
import { RegisterProfileNotificationsBody } from "@workspace/api-zod";
import { SendTestProfileNotificationBody } from "@workspace/api-zod";
import { UploadProfilePhotoBody } from "@workspace/api-zod";
import {
  failureReasonLabel,
  hub2Environment,
  hub2Headers,
  paymentStatusFromHub2,
} from "./premium";
import { requireAuth, requireNonDemo } from "../middleware/auth";
import {
  isFirebasePushToken,
  sendTestPushNotification,
} from "../lib/pushNotifications";
import { createPremiumTrial, getPremiumTrialStatus } from "../lib/premiumTrial";

const router: IRouter = Router();
router.use("/profile", requireAuth);
const PROFESSION_CHANGE_AMOUNT = 2000;
const PROFESSION_CHANGE_WINDOW_MS = 30 * 24 * 60 * 60 * 1000;
// Temporary test-period switch. Set PROFESSION_CHANGE_TEST_MODE=false to restore
// the monthly limit and Hub2 payment requirement.
const PROFESSION_CHANGE_TEST_MODE =
  process.env.PROFESSION_CHANGE_TEST_MODE?.toLowerCase() !== "false";

// ---------------------------------------------------------------------------
// In-memory delete-token store
//
// Tokens are generated by POST /profile/delete-token and consumed by
// DELETE /profile. They are:
//   - Short-lived (15 minutes)
//   - Single-use (removed from the store on first use)
//   - Never embedded in GET /profile responses
//
// The token maps to the exact profile id it was issued for, so DELETE
// always resolves the profile from the token's stored identity rather
// than an unscoped limit(1) query.
// ---------------------------------------------------------------------------
const DELETE_TOKEN_TTL_MS = 15 * 60 * 1000;

interface StoredDeleteToken {
  profileId: number;
  expiresAt: number;
}

const deleteTokenStore = new Map<string, StoredDeleteToken>();

function issueDeleteToken(profileId: number): { token: string; expiresAt: Date } {
  // Purge expired entries opportunistically
  const now = Date.now();
  for (const [key, entry] of deleteTokenStore) {
    if (entry.expiresAt <= now) deleteTokenStore.delete(key);
  }
  const token = randomBytes(32).toString("hex");
  const expiresAt = new Date(now + DELETE_TOKEN_TTL_MS);
  deleteTokenStore.set(token, { profileId, expiresAt: expiresAt.getTime() });
  return { token, expiresAt };
}

function consumeDeleteToken(token: string): number | null {
  const entry = deleteTokenStore.get(token);
  if (!entry) return null;
  deleteTokenStore.delete(token); // Single-use: always remove after first lookup
  if (entry.expiresAt <= Date.now()) return null;
  return entry.profileId;
}

function professionChangeAvailableAt(lastChangeAt: Date | null): Date | null {
  if (!lastChangeAt) return null;
  const availableAt = new Date(lastChangeAt.getTime() + PROFESSION_CHANGE_WINDOW_MS);
  return availableAt.getTime() <= Date.now() ? null : availableAt;
}

function isPlaceholderProfileName(value: string | null | undefined): boolean {
  const normalized = value?.trim().toLocaleLowerCase("fr-FR") ?? "";
  return !normalized || normalized === "nouveau candidat" || normalized === "utilisateur" || normalized === "new candidate";
}

function fullNameFromCvAnalysis(value: unknown): string | null {
  if (!value || typeof value !== "object") return null;
  const fullName = (value as { fullName?: unknown }).fullName;
  return typeof fullName === "string" && !isPlaceholderProfileName(fullName) ? fullName.trim() : null;
}

function makeProfessionPurchaseReference(profileId: number): string {
  return `JOBAGOGO_PROFESSION_${profileId}_${Date.now()}_${Math.random().toString(36).slice(2, 8)}`;
}

function professionPaymentJson(
  payment: typeof professionChangePaymentsTable.$inferSelect,
  checkoutUrl: string,
) {
  const environment = hub2Environment();
  return {
    id: payment.id,
    amount: payment.amount,
    status: payment.status,
    checkoutUrl,
    purchaseReference: payment.purchaseReference,
    paymentLinkExpiresAt: payment.paymentLinkExpiresAt?.toISOString() ?? null,
    environment,
    sandboxTestPhone: environment === "sandbox" ? "00000001" : null,
  };
}

function profileShape(profile: typeof profilesTable.$inferSelect, isDemo = false) {
  const premiumStatus = getPremiumTrialStatus(profile);

  return {
    id: profile.id,
    name: profile.name,
    title: profile.title,
    skills: profile.skills,
    experienceYears: profile.experienceYears,
    location: profile.location,
    remote: profile.remote,
    jobTypes: profile.jobTypes,
    salaryMin: profile.salaryMin,
    salaryMax: profile.salaryMax,
    bio: profile.bio,
    email: profile.email,
    linkedin: profile.linkedin,
    phone: profile.phone,
    diploma: profile.diploma,
    lastJobTitle: profile.lastJobTitle,
    profilePhoto: profile.profilePhoto,
    notificationsEnabled: profile.notificationsEnabled,
    lastProfessionChangeAt: profile.lastProfessionChangeAt?.toISOString() ?? null,
    professionChangeAvailableAt: PROFESSION_CHANGE_TEST_MODE
      ? null
      : professionChangeAvailableAt(profile.lastProfessionChangeAt)?.toISOString() ?? null,
    isPremium: premiumStatus.isPremium,
    premiumUntil: profile.premiumUntil?.toISOString() ?? null,
    premiumTrialUntil: profile.premiumTrialUntil?.toISOString() ?? null,
    isPremiumTrial: premiumStatus.isPremiumTrial,
    isPremiumTrialExpired: premiumStatus.isPremiumTrialExpired,
    updatedAt: profile.updatedAt.toISOString(),
    // Whether a device credential (SHA-256 hash) is stored for this account.
    // The mobile client uses this to decide whether to show the in-app delete
    // flow or a "contact support" message for accounts created before this
    // feature was introduced.
    isDeviceRegistered: profile.deviceSecretHash !== null,
    isDemo,
  };
}

router.get("/profile", async (req, res): Promise<void> => {
  const profiles = await db.select().from(profilesTable)
    .where(eq(profilesTable.id, req.auth!.profileId))
    .limit(1);
  if (!profiles[0]) {
    res.status(404).json({ error: "No profile found" });
    return;
  }
  const [latestCv] = await db.select({
    id: cvDocumentsTable.id,
    fileName: cvDocumentsTable.fileName,
    status: cvDocumentsTable.status,
    analysis: cvDocumentsTable.analysis,
    createdAt: cvDocumentsTable.createdAt,
  }).from(cvDocumentsTable)
    .where(eq(cvDocumentsTable.profileId, profiles[0].id))
    .orderBy(desc(cvDocumentsTable.createdAt))
    .limit(1);
  let profileForResponse = profiles[0];
  const inferredName = isPlaceholderProfileName(profileForResponse.name)
    ? fullNameFromCvAnalysis(latestCv?.analysis)
    : null;
  if (inferredName) {
    const [updatedProfile] = await db.update(profilesTable)
      .set({ name: inferredName })
      .where(eq(profilesTable.id, profileForResponse.id))
      .returning();
    profileForResponse = updatedProfile ?? { ...profileForResponse, name: inferredName };
  }

  res.json({ ...profileShape(profileForResponse, req.auth!.isDemo), latestCv: latestCv ? {
    ...latestCv,
    createdAt: latestCv.createdAt.toISOString(),
  } : null });
});

router.post("/profile", requireNonDemo, async (req, res): Promise<void> => {
  const parsed = UpsertProfileBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const existing = await db.select().from(profilesTable)
    .where(eq(profilesTable.id, req.auth!.profileId))
    .limit(1);

  let profile;
  if (existing[0]) {
    // UPDATE path: never touch deviceSecretHash here.
    // The deletion credential is established at profile creation and can only
    // be registered afterwards via POST /profile/register-device, which only
    // works when no hash is stored yet. This prevents an unauthenticated
    // caller from overwriting the credential via a profile update.
    const currentProfile = existing[0];
    const requestedTitle = parsed.data.title?.trim() ?? "";
    const currentTitle = currentProfile.title?.trim() ?? "";
    const professionChanged = requestedTitle !== currentTitle;
    const availableAt = professionChangeAvailableAt(currentProfile.lastProfessionChangeAt);

    if (professionChanged && availableAt && !PROFESSION_CHANGE_TEST_MODE) {
      res.status(409).json({
        error: "Changer de métier avant la fin du mois nécessite un paiement de 2 000 XOF.",
        code: "PROFESSION_CHANGE_PAYMENT_REQUIRED",
        professionChangeAvailableAt: availableAt.toISOString(),
        amount: PROFESSION_CHANGE_AMOUNT,
      });
      return;
    }

    const updated = await db
      .update(profilesTable)
      .set({
        ...parsed.data,
        email: currentProfile.email,
        emailVerifiedAt: currentProfile.emailVerifiedAt,
        skills: parsed.data.skills ?? [],
        jobTypes: parsed.data.jobTypes ?? [],
        ...(professionChanged ? { lastProfessionChangeAt: new Date() } : {}),
      })
      .where(eq(profilesTable.id, existing[0].id))
      .returning();
    profile = updated[0];
  } else {
    const premiumTrial = createPremiumTrial();
    const inserted = await db
      .insert(profilesTable)
      .values({
        ...parsed.data,
        email: req.auth!.email,
        emailVerifiedAt: new Date(),
        ...premiumTrial,
        skills: parsed.data.skills ?? [],
        jobTypes: parsed.data.jobTypes ?? [],
      })
      .returning();
    profile = inserted[0];
  }

  res.json(profileShape(profile));
});

router.post("/profile/photo", requireNonDemo, async (req, res): Promise<void> => {
  const parsed = UploadProfilePhotoBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [profile] = await db
    .update(profilesTable)
    .set({ profilePhoto: parsed.data.profilePhoto })
    .where(eq(profilesTable.id, req.auth!.profileId))
    .returning();

  if (!profile) {
    res.status(404).json({ error: "No profile found" });
    return;
  }

  res.json(profileShape(profile));
});

router.delete("/profile/photo", requireNonDemo, async (req, res): Promise<void> => {
  const [profile] = await db
    .update(profilesTable)
    .set({ profilePhoto: null })
    .where(eq(profilesTable.id, req.auth!.profileId))
    .returning();

  if (!profile) {
    res.status(404).json({ error: "No profile found" });
    return;
  }

  res.json(profileShape(profile));
});

router.post("/profile/notifications", requireNonDemo, async (req, res): Promise<void> => {
  const parsed = RegisterProfileNotificationsBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  const [profile] = await db.select().from(profilesTable)
    .where(eq(profilesTable.id, req.auth!.profileId))
    .limit(1);
  if (!profile) {
    res.status(404).json({ error: "No profile found" });
    return;
  }

  const enabled = parsed.data.enabled && Boolean(parsed.data.token);
  if (enabled && (!parsed.data.token || !isFirebasePushToken(parsed.data.token))) {
    res.status(400).json({
      error: "Le token Firebase est invalide. Réessaie depuis une build mobile récente.",
    });
    return;
  }

  if (enabled && parsed.data.token) {
    await db
      .insert(pushDevicesTable)
      .values({
        profileId: profile.id,
        token: parsed.data.token,
        platform: parsed.data.platform ?? "android",
        enabled: true,
        lastSeenAt: new Date(),
      })
      .onConflictDoUpdate({
        target: pushDevicesTable.token,
        set: {
          profileId: profile.id,
          platform: parsed.data.platform ?? "android",
          enabled: true,
          lastSeenAt: new Date(),
          updatedAt: new Date(),
        },
      });
  } else if (parsed.data.token) {
    await db
      .update(pushDevicesTable)
      .set({ enabled: false, updatedAt: new Date() })
      .where(
        and(
          eq(pushDevicesTable.profileId, profile.id),
          eq(pushDevicesTable.token, parsed.data.token),
        ),
      );
  } else {
    await db
      .update(pushDevicesTable)
      .set({ enabled: false, updatedAt: new Date() })
      .where(eq(pushDevicesTable.profileId, profile.id));
  }

  const [activeDevice] = await db
    .select({ id: pushDevicesTable.id })
    .from(pushDevicesTable)
    .where(
      and(
        eq(pushDevicesTable.profileId, profile.id),
        eq(pushDevicesTable.enabled, true),
      ),
    )
    .limit(1);

  await db
    .update(profilesTable)
    .set({ notificationsEnabled: Boolean(activeDevice) })
    .where(eq(profilesTable.id, profile.id));

  res.json({ enabled: Boolean(activeDevice), provider: "firebase" as const });
});

router.post("/profile/notifications/test", requireNonDemo, async (req, res): Promise<void> => {
  const parsed = SendTestProfileNotificationBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }

  try {
    const result = await sendTestPushNotification(req.auth!.profileId, parsed.data.jobId);
    res.json(result);
  } catch (error) {
    const statusCode =
      typeof (error as { statusCode?: unknown }).statusCode === "number"
        ? (error as { statusCode: number }).statusCode
        : 502;
    res.status(statusCode).json({
      error: error instanceof Error ? error.message : "La notification de test n'a pas pu être envoyée.",
    });
  }
});

router.post("/profile/profession-change/payments", requireNonDemo, async (req, res): Promise<void> => {
  if (PROFESSION_CHANGE_TEST_MODE) {
    res.status(400).json({
      error: "Le changement de métier est gratuit pendant la période de test.",
    });
    return;
  }

  const requestedTitle = typeof req.body?.title === "string" ? req.body.title.trim() : "";
  if (!requestedTitle) {
    res.status(400).json({ error: "Le nouveau métier est obligatoire." });
    return;
  }

  const [profile] = await db.select().from(profilesTable)
    .where(eq(profilesTable.id, req.auth!.profileId))
    .limit(1);
  if (!profile) {
    res.status(404).json({ error: "Profil introuvable" });
    return;
  }

  const currentTitle = profile.title?.trim() ?? "";
  const availableAt = professionChangeAvailableAt(profile.lastProfessionChangeAt);
  if (!currentTitle || requestedTitle === currentTitle || !availableAt) {
    res.status(400).json({ error: "Ce changement de métier ne nécessite pas de paiement." });
    return;
  }

  const purchaseReference = makeProfessionPurchaseReference(profile.id);
  const expirationDate = new Date(Date.now() + 30 * 60 * 1000);
  let hub2Response: Response;
  try {
    hub2Response = await fetch("https://api.hub2.io/payment-links", {
      method: "POST",
      headers: hub2Headers(),
      body: JSON.stringify({
        purchaseReference,
        description: `Changement de métier Jobagogo : ${requestedTitle}`,
        amount: PROFESSION_CHANGE_AMOUNT,
        currency: "XOF",
        country: process.env["HUB2_COUNTRY"] ?? "CI",
        type: "single_use",
        expirationDate: expirationDate.toISOString(),
        paymentMethods: ["mobile_money"],
        providers: ["Orange", "MTN", "Moov", "Wave"],
      }),
    });
  } catch (error) {
    req.log.error({ err: error }, "Hub2 profession change payment request failed");
    res.status(502).json({ error: "Impossible de joindre Hub2" });
    return;
  }

  if (!hub2Response.ok) {
    req.log.error({
      status: hub2Response.status,
      body: await hub2Response.text(),
    }, "Hub2 rejected profession change payment");
    res.status(502).json({ error: "Hub2 a refusé la création du paiement" });
    return;
  }

  const hub2PaymentLink = await hub2Response.json() as { id?: string; url?: string };
  if (!hub2PaymentLink.id || !hub2PaymentLink.url) {
    res.status(502).json({ error: "Réponse Hub2 invalide" });
    return;
  }

  const [payment] = await db.insert(professionChangePaymentsTable).values({
    profileId: profile.id,
    hub2PaymentLinkId: hub2PaymentLink.id,
    purchaseReference,
    requestedTitle,
    amount: PROFESSION_CHANGE_AMOUNT,
    paymentLinkExpiresAt: expirationDate,
  }).returning();

  res.status(201).json(professionPaymentJson(payment, hub2PaymentLink.url));
});

router.get("/profile/profession-change/payments/:paymentId/status", requireNonDemo, async (req, res): Promise<void> => {
  const paymentId = Number(req.params["paymentId"]);
  if (!Number.isInteger(paymentId)) {
    res.status(400).json({ error: "Identifiant de paiement invalide" });
    return;
  }

  const [payment] = await db.select()
    .from(professionChangePaymentsTable)
    .where(and(
      eq(professionChangePaymentsTable.id, paymentId),
      eq(professionChangePaymentsTable.profileId, req.auth!.profileId),
    ))
    .limit(1);
  if (!payment) {
    res.status(404).json({ error: "Paiement introuvable" });
    return;
  }

  let hub2Response: Response;
  try {
    hub2Response = await fetch(
      `https://api.hub2.io/payment-links/${encodeURIComponent(payment.hub2PaymentLinkId)}`,
      { headers: hub2Headers() },
    );
  } catch (error) {
    req.log.error({ err: error, paymentId }, "Hub2 profession change status request failed");
    res.status(502).json({ error: "Impossible de vérifier le paiement auprès de Hub2" });
    return;
  }

  if (!hub2Response.ok) {
    res.status(502).json({ error: "Hub2 n'a pas pu vérifier le paiement" });
    return;
  }

  const hub2PaymentLink = await hub2Response.json();
  const { status, failureReason } = paymentStatusFromHub2(hub2PaymentLink);
  const completed = status === "completed";

  if (completed && payment.status !== "completed") {
    await db.update(professionChangePaymentsTable)
      .set({ status: "completed" })
      .where(eq(professionChangePaymentsTable.id, payment.id));
    await db.update(profilesTable)
      .set({ title: payment.requestedTitle, lastProfessionChangeAt: new Date() })
      .where(eq(profilesTable.id, payment.profileId));
  } else if (!completed && status !== payment.status) {
    await db.update(professionChangePaymentsTable)
      .set({ status })
      .where(eq(professionChangePaymentsTable.id, payment.id));
  }

  res.json({
    id: payment.id,
    status: completed ? "completed" : status,
    title: payment.requestedTitle,
    failureReason: status === "failed" ? failureReasonLabel(failureReason) : null,
    message: completed
      ? "Paiement confirmé. Le métier a été changé."
      : status === "failed"
        ? failureReasonLabel(failureReason)
        : "Paiement en attente de confirmation.",
  });
});

// POST /profile/delete-token — issues a short-lived, single-use token that
// the client must present to DELETE /profile. The token is stored server-side
// and is never embedded in the GET /profile response.
//
// Primary authentication is the verified email session. The old device secret
// remains on clients only for compatibility with older builds.
//
router.post("/profile/delete-token", requireNonDemo, async (req, res): Promise<void> => {
  const [profile] = await db.select().from(profilesTable)
    .where(eq(profilesTable.id, req.auth!.profileId))
    .limit(1);
  if (!profile) {
    res.status(404).json({ error: "No profile found" });
    return;
  }

  const { token, expiresAt } = issueDeleteToken(profile.id);
  res.json({ token, expiresAt: expiresAt.toISOString() });
});

// DELETE /profile — permanently purges the profile and all associated data.
// The authenticated session remains in Authorization; the short-lived
// confirmation token is sent separately so the auth middleware can still
// identify the account. The token is single-use and must belong to the same
// profile as the session.
router.delete("/profile", requireNonDemo, async (req, res): Promise<void> => {
  const rawToken = req.headers["x-delete-token"];
  if (typeof rawToken !== "string" || rawToken.length === 0) {
    res.status(401).json({ error: "Token de confirmation requis." });
    return;
  }
  const profileId = consumeDeleteToken(rawToken);
  if (!profileId || profileId !== req.auth!.profileId) {
    res.status(401).json({ error: "Token de suppression invalide, expiré ou déjà utilisé." });
    return;
  }

  const [profile] = await db
    .select()
    .from(profilesTable)
    .where(eq(profilesTable.id, profileId))
    .limit(1);
  if (!profile) {
    res.status(404).json({ error: "No profile found" });
    return;
  }

  // Purge all profile data in a single transaction so a partial failure
  // never leaves the account in an inconsistent state.
  await db.transaction(async (tx) => {
    await tx.delete(cvDocumentsTable).where(eq(cvDocumentsTable.profileId, profile.id));
    await tx.delete(premiumPaymentsTable).where(eq(premiumPaymentsTable.profileId, profile.id));
    await tx.delete(professionChangePaymentsTable).where(eq(professionChangePaymentsTable.profileId, profile.id));
    await tx.delete(savedJobsTable).where(eq(savedJobsTable.profileId, profile.id));
    await tx.delete(profilesTable).where(eq(profilesTable.id, profile.id));
  });

  res.status(204).end();
});

export default router;