# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JobMatch API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get candidate profile
      description: Returns the stored candidate profile
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateProfile"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProfile
      tags: [profile]
      summary: Delete the current profile and all associated data
      description: >
        Permanently deletes the profile, CVs, saved jobs, and payment records.
        Requires the authenticated email session and a short-lived confirmation
        token obtained from POST /profile/delete-token. The confirmation token
        is single-use and expires after 15 minutes.
      parameters:
        - name: X-Delete-Token
          in: header
          required: true
          description: "One-time confirmation token obtained from POST /profile/delete-token"
          schema:
            type: string
      responses:
        "204":
          description: Profile and all associated data deleted
        "401":
          description: Missing or invalid delete token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateProfile"
        "409":
          description: Profession change requires payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessionChangeRequiredError"
  /profile/delete-token:
    post:
      operationId: createProfileDeleteToken
      tags: [profile]
      summary: Request a short-lived token to delete the current profile
      description: >
        Issues a one-time, 15-minute confirmation token that must be presented
        in X-Delete-Token to DELETE /profile. The token is stored server-side
        and is never returned by GET /profile. The current authenticated email
        session is required.
      responses:
        "200":
          description: Delete token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeleteToken"
        "401":
          description: Missing, invalid, or unregistered device secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Request originated from a web browser (Origin header present)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile/photo:
    post:
      operationId: uploadProfilePhoto
      tags: [profile]
      summary: Add or replace the current profile photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfilePhotoInput"
      responses:
        "200":
          description: Profile photo saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateProfile"
        "400":
          description: Invalid or oversized photo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProfilePhoto
      tags: [profile]
      summary: Remove the current profile photo
      responses:
        "200":
          description: Profile photo removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateProfile"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile/notifications:
    post:
      operationId: registerProfileNotifications
      tags: [profile]
      summary: Enable or disable Firebase notifications for the current profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationRegistrationInput"
      responses:
        "200":
          description: Notification preference saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationRegistration"
        "400":
          description: Invalid notification registration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile/notifications/test:
    post:
      operationId: sendTestProfileNotification
      tags: [profile]
      summary: Send an immediate test notification for an existing job
      description: >
        Sends a notification to all Firebase devices registered for the current profile
        for the selected existing job. This bypasses the scheduled scrape and
        normal match threshold, but still requires notifications to be enabled
        and a valid push token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationTestInput"
      responses:
        "200":
          description: Test notification accepted by Firebase Cloud Messaging
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationTest"
        "400":
          description: Notifications are not enabled or request is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile or job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Firebase rejected the notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /profile/profession-change/payments:
    post:
      operationId: createProfessionChangePayment
      tags: [profile]
      summary: Create a Hub2 payment link to change profession before the monthly limit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfessionChangePaymentInput"
      responses:
        "201":
          description: Profession change payment link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessionChangePayment"
        "400":
          description: Profession change is not eligible for payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Hub2 unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile/profession-change/payments/{paymentId}/status:
    get:
      operationId: getProfessionChangePaymentStatus
      tags: [profile]
      summary: Verify a profession change payment with Hub2
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profession change payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessionChangePaymentStatus"
        "404":
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    AccountDeleteToken:
      type: object
      required: [token, expiresAt]
      properties:
        token:
          type: string
          description: One-time Bearer token to present to DELETE /profile
        expiresAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp after which the token expires
    CandidateProfile:
      type: object
      required: [id, name, skills, experienceYears, location, jobTypes, updatedAt, isDemo]
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        location:
          type: string
        remote:
          type: boolean
        jobTypes:
          type: array
          items:
            type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        bio:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        diploma:
          type: ["string", "null"]
        lastJobTitle:
          type: ["string", "null"]
        profilePhoto:
          type: ["string", "null"]
          description: Private compressed image data for the authenticated user's profile
        lastProfessionChangeAt:
          type: ["string", "null"]
        professionChangeAvailableAt:
          type: ["string", "null"]
        notificationsEnabled:
          type: boolean
        isPremium:
          type: boolean
        premiumUntil:
          type: ["string", "null"]
        premiumTrialUntil:
          type: ["string", "null"]
          format: date-time
          description: End timestamp for the one-time free Premium trial, when applicable
        isPremiumTrial:
          type: boolean
          description: Whether this profile currently has Premium through its free trial
        isPremiumTrialExpired:
          type: boolean
          description: Whether the free trial expired without a paid Premium extension
        latestCv:
          allOf:
            - $ref: "#/components/schemas/CvDocument"
          nullable: true
        updatedAt:
          type: string
        isDeviceRegistered:
          type: boolean
          description: >
            True when a device-credential hash is stored for this account. The
            mobile client uses this to determine whether the in-app delete flow
            is available or whether the user should be directed to a support
            contact for legacy accounts created before the credential feature.
        isDemo:
          type: boolean
          description: Legacy session marker; retired demo sessions are no longer accepted
    CandidateProfileInput:
      type: object
      required: [name, skills, experienceYears, location, jobTypes]
      properties:
        name:
          type: string
        title:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        location:
          type: string
        remote:
          type: boolean
        jobTypes:
          type: array
          items:
            type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        bio:
          type: string
        email:
          type: string
        linkedin:
          type: string
        phone:
          type: string
        diploma:
          type: string
        lastJobTitle:
          type: string
    UpdateProfilePhotoInput:
      type: object
      required: [profilePhoto]
      properties:
        profilePhoto:
          type: string
          maxLength: 1600000
          pattern: "^data:image/(jpeg|png|webp);base64,"
          description: Compressed image as a data URL
...
    ProfessionChangeRequiredError:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            code:
              type: string
              enum: [PROFESSION_CHANGE_PAYMENT_REQUIRED]
            professionChangeAvailableAt:
              type: string
            amount:
              type: integer
              example: 2000
    NotificationRegistrationInput:
      type: object
      required: [enabled]
      properties:
        token:
          type: ["string", "null"]
          maxLength: 4096
        enabled:
          type: boolean
        platform:
          type: string
          enum: [ios, android]
...